# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"EasyScraper API","summary":"API for EasyScraper","version":"1.0.0"},"paths":{"/user/signup":{"post":{"tags":["auth"],"summary":"Sign Up","operationId":"sign_up_user_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"409":{"description":"user already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAccessToken"}}}},"401":{"description":"wrong username or password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects":{"post":{"tags":["projects"],"summary":"Create Project","operationId":"create_project_projects_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["projects"],"summary":"List Projects","operationId":"list_projects_projects_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}":{"get":{"tags":["projects"],"summary":"Get Project","operationId":"get_project_projects__project_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["projects"],"summary":"Put Project","operationId":"put_project_projects__project_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["projects"],"summary":"Delete Project","operationId":"delete_project_projects__project_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/page_templates":{"post":{"tags":["page templates"],"summary":"Create Page Template","operationId":"create_page_template_projects__project_id__page_templates_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplateCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["page templates"],"summary":"List Page Templates","operationId":"list_page_templates_projects__project_id__page_templates_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplateListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/page_templates/{page_template_id}":{"get":{"tags":["page templates"],"summary":"Get Page Template","operationId":"get_page_template_projects__project_id__page_templates__page_template_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplate"}}}},"404":{"description":"PageTemplate not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["page templates"],"summary":"Put Page Template","operationId":"put_page_template_projects__project_id__page_templates__page_template_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplateUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageTemplate"}}}},"404":{"description":"PageTemplate not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["page templates"],"summary":"Delete Page Template","operationId":"delete_page_template_projects__project_id__page_templates__page_template_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdResponse"}}}},"404":{"description":"PageTemplate not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/page_templates/{page_template_id}/scrape_rules":{"post":{"tags":["scrape rules"],"summary":"Create Scrape Rule","operationId":"create_scrape_rule_projects__project_id__page_templates__page_template_id__scrape_rules_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRuleCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRule"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["scrape rules"],"summary":"List Scrape Rules","operationId":"list_scrape_rules_projects__project_id__page_templates__page_template_id__scrape_rules_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRuleListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/page_templates/{page_template_id}/scrape_rules/{scrape_rule_id}":{"get":{"tags":["scrape rules"],"summary":"Get Scrape Rule","operationId":"get_scrape_rule_projects__project_id__page_templates__page_template_id__scrape_rules__scrape_rule_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}},{"name":"scrape_rule_id","in":"path","required":true,"schema":{"title":"Scrape Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRule"}}}},"404":{"description":"ScrapeRule not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["scrape rules"],"summary":"Update Scrape Rule","operationId":"update_scrape_rule_projects__project_id__page_templates__page_template_id__scrape_rules__scrape_rule_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}},{"name":"scrape_rule_id","in":"path","required":true,"schema":{"title":"Scrape Rule Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRuleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRule"}}}},"404":{"description":"ScrapeRule not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["scrape rules"],"summary":"Delete Scrape Rule","operationId":"delete_scrape_rule_projects__project_id__page_templates__page_template_id__scrape_rules__scrape_rule_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}},{"name":"page_template_id","in":"path","required":true,"schema":{"title":"Page Template Id"}},{"name":"scrape_rule_id","in":"path","required":true,"schema":{"title":"Scrape Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdResponse"}}}},"404":{"description":"ScrapeRule not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FastAPIError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["miscellaneous"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/celery-test/{project_id}":{"get":{"tags":["miscellaneous"],"summary":"Celery Test","operationId":"celery_test_celery_test__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FastAPIError":{"properties":{"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"FastAPIError"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IdResponse":{"properties":{"id":{"type":"string","title":"Id"}},"type":"object","required":["id"],"title":"IdResponse"},"PageTemplate":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"output_type":{"type":"string","enum":["PAGE_SOURCE","LEAF"],"title":"Output Type"},"scraper":{"type":"string","enum":["XPATH_SELECTOR","CSS_SELECTOR","AUTO_SCRAPER","AI_SCRAPER"],"title":"Scraper"},"output_page_template_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Page Template Id"},"project_id":{"type":"string","title":"Project Id"}},"type":"object","required":["name","output_type","scraper","project_id"],"title":"PageTemplate"},"PageTemplateCreate":{"properties":{"name":{"type":"string","title":"Name"},"output_type":{"type":"string","enum":["PAGE_SOURCE","LEAF"],"title":"Output Type"},"scraper":{"type":"string","enum":["XPATH_SELECTOR","CSS_SELECTOR","AUTO_SCRAPER","AI_SCRAPER"],"title":"Scraper"},"output_page_template_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Page Template Id"}},"type":"object","required":["name","output_type","scraper"],"title":"PageTemplateCreate"},"PageTemplateListResponse":{"properties":{"page_templates":{"items":{"$ref":"#/components/schemas/PageTemplate"},"type":"array","title":"Page Templates"},"paging":{"$ref":"#/components/schemas/Paging"}},"type":"object","required":["page_templates","paging"],"title":"PageTemplateListResponse"},"PageTemplateUpdate":{"properties":{"name":{"type":"string","title":"Name"},"output_type":{"type":"string","enum":["PAGE_SOURCE","LEAF"],"title":"Output Type"},"scraper":{"type":"string","enum":["XPATH_SELECTOR","CSS_SELECTOR","AUTO_SCRAPER","AI_SCRAPER"],"title":"Scraper"},"output_page_template_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Page Template Id"}},"type":"object","title":"PageTemplateUpdate"},"Paging":{"properties":{"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["skip","limit"],"title":"Paging"},"Project":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"sleep_seconds_between_page_scrape":{"type":"integer","title":"Sleep Seconds Between Page Scrape","default":3},"user_id":{"type":"string","title":"User Id"}},"type":"object","required":["name","user_id"],"title":"Project"},"ProjectCreate":{"properties":{"name":{"type":"string","title":"Name"},"sleep_seconds_between_page_scrape":{"type":"integer","title":"Sleep Seconds Between Page Scrape","default":3}},"type":"object","required":["name"],"title":"ProjectCreate"},"ProjectUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"sleep_seconds_between_page_scrape":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sleep Seconds Between Page Scrape"}},"type":"object","title":"ProjectUpdate"},"ScrapeRule":{"properties":{"id":{"type":"string","title":"Id"},"alias":{"type":"string","title":"Alias"},"type":{"type":"string","enum":["SINGLE","MULTI"],"title":"Type"},"value":{"type":"string","title":"Value"},"page_template_id":{"type":"string","title":"Page Template Id"}},"type":"object","required":["alias","type"],"title":"ScrapeRule"},"ScrapeRuleCreate":{"properties":{"alias":{"type":"string","title":"Alias"},"type":{"type":"string","enum":["SINGLE","MULTI"],"title":"Type"},"value":{"type":"string","title":"Value"}},"type":"object","required":["alias","type"],"title":"ScrapeRuleCreate"},"ScrapeRuleListResponse":{"properties":{"scrape_rules":{"items":{"$ref":"#/components/schemas/ScrapeRule"},"type":"array","title":"Scrape Rules"},"paging":{"$ref":"#/components/schemas/Paging"}},"type":"object","required":["scrape_rules","paging"],"title":"ScrapeRuleListResponse"},"ScrapeRuleUpdate":{"properties":{"alias":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Alias"},"type":{"type":"string","enum":["SINGLE","MULTI"],"title":"Type"},"value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Value"}},"type":"object","title":"ScrapeRuleUpdate"},"UserAccessToken":{"properties":{"username":{"type":"string","title":"Username"},"token":{"type":"string","title":"Token"}},"type":"object","required":["username","token"],"title":"UserAccessToken"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":20,"minLength":6,"title":"Username"},"password":{"type":"string","maxLength":20,"minLength":6,"title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserLogin":{"properties":{"username":{"type":"string","maxLength":20,"minLength":6,"title":"Username"},"password":{"type":"string","maxLength":20,"minLength":6,"title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin"},"UserOut":{"properties":{"id":{"type":"string","title":"Id"},"username":{"type":"string","title":"Username"}},"type":"object","required":["id","username"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"auth","description":"Authentication endpoints for user login and registration"},{"name":"projects","description":"Endpoints for managing scraping projects and their configurations"},{"name":"page templates","description":"Endpoints for managing page templates"},{"name":"scrape rules","description":"Endpoints for managing scrape rules"},{"name":"miscellaneous","description":"Endpoints for managing scraping projects and their configurations"}]}